fix tip percentage not using decimal points

+++ Tasks/Tasks_03/03_Tip_Calculator_MVC/src/scripts/app.ts
@@ -24,7 +24,7 @@ calculateTipForm.addEventListener("submit", (event) => {
     event.preventDefault();
     const tipModel = new TipModel(
         parseInt(amountOfPersonsInputField.value),
-        parseInt(tipPercentageInputField.value),
+        parseFloat(tipPercentageInputField.value),
         parseFloat(grossAmountInputField.value)
     );
